Add GetSongIndex method and change AddSong return type
Implemented a method to get the index of a specific song in the SongDatabase. Also changed the return type of the AddSong method from void to size_t. These changes will allow other parts of the application to easily find the location of a newly added song or an existing one in the database.