Add live song updates to Now Playing panel
This change responds to a need for displaying live song changes in the 'Now Playing' panel. An 'Update()' function was added to `NowPlayingPanel.cpp`, taking `wxUpdateUiEvent`. Further, a new member, `mEventReceiver`, was included in `NowPlayingPanel`. It is initialized in the `NowPlayingPanel` constructor and used in `Update()` method to receive song change events and correspondingly update the song title label. The class `NowPlayingPanel` now also includes `ChannelReceiver` and `Playlist` headers for the newly incorporated functionalities.