Add functionality to remove songs from playlist
Code changes were done in MusicPanel.cpp and MusicPanel.hpp. The primary purpose of these modifications is to add the ability for the user to remove songs directly from the playlist. This capability aids in the better management by the user of their song collection in the playlist. In addition to this, minor refactoring such as cleaning up the tab indentations and setting up more consistent naming conventions for some identifiers was done for better coding standard adherence.