"Add highlight to currently playing song in MusicPanel"
A new feature was added to the UI for better user experience. Now when a song is changed in the playlist, not only will the previous song be removed from the playlist view, but the currently playing song will also be highlighted with a distinct background color. This makes it easier for users to identify currently playing track.