Update NowPlayingPanel to show "No Song Playing" when playback ends
Added a new condition in NowPlayingPanel.cpp to check for PlaybackEndedEvent. Now, when the playback ends, the 'Now Playing' section in the UI will show "No Song Playing" instead of retaining the title of the last played song. This gives a clearer indication to users that the player is not currently playing any