Add SetTitle function and JSON serialization for Song model
A SetTitle function was added to the Song model, allowing for the title of a Song to be changed after creation. Moreover, a ToJSON function was added to provide serialization of a Song object to JSON format. These additions are crucial for the functionality of the program since it allows for songs to be edited and object states to be stored and retrieved.