Added feature to rename songs in Music Panel
A new feature has been introduced to rename songs within the Music Panel. This change involved adding a new "Rename" button to the existing ones for Add, Enqueue, and Remove Songs. A new event listener has also been added for this renaming functionality. The related changes have also been made to the SongDatabase to track and make changes to the new song name.